Fixed AES encryption/decryption
Previously each encryption appended an authentication tag and a nonce to the encrypted payload.
This made it impossible to know the exact size of the payload for decryption.
Now the tag is omitted and only one nonce is used per CircuitOpaque message, which makes the payload have a constant size.